AND-3770 Moved entities to another package.
