+++ blockchain/src/main/java/com/tangem/blockchain/common/assembly/WalletManagerAssembly.kt
@@ -1,0 +1,13 @@
+package com.tangem.blockchain.common.assembly
+
+import com.tangem.blockchain.common.*
+
+/**
+ * Creates WalletManager instance
+ */
+abstract class WalletManagerAssembly<out T : WalletManager> {
+
+    abstract fun make(input: WalletManagerAssemblyInput): T
+
+}
+
