+++ blockchain/src/main/java/com/tangem/blockchain/common/assembly/WalletManagerAssemblyInput.kt
@@ -3,15 +3,11 @@ package com.tangem.blockchain.common.assembly
 import com.tangem.blockchain.common.BlockchainSdkConfig
 import com.tangem.blockchain.common.Token
 import com.tangem.blockchain.common.Wallet
-import com.tangem.blockchain.common.WalletManager
 import com.tangem.common.card.EllipticCurve
 
-abstract class WalletManagerAssembly<out T : WalletManager> {
-
-    abstract fun make(input: WalletManagerAssemblyInput): T
-
-}
-
+/**
+ * Data needed for WalletManager creation
+ */
 class WalletManagerAssemblyInput(
     val wallet: Wallet,
     val config: BlockchainSdkConfig,
