+++ blockchain/src/main/java/com/tangem/blockchain/common/assembly/impl/DucatusWalletManagerAssembly.kt
@@ -1,8 +1,10 @@
-package com.tangem.blockchain.common.assembly
+package com.tangem.blockchain.common.assembly.impl
 
 import com.tangem.blockchain.blockchains.bitcoin.BitcoinTransactionBuilder
 import com.tangem.blockchain.blockchains.ducatus.DucatusWalletManager
 import com.tangem.blockchain.blockchains.ducatus.network.DucatusNetworkService
+import com.tangem.blockchain.common.assembly.WalletManagerAssembly
+import com.tangem.blockchain.common.assembly.WalletManagerAssemblyInput
 
 object DucatusWalletManagerAssembly : WalletManagerAssembly<DucatusWalletManager>() {
