+++ blockchain/src/main/java/com/tangem/blockchain/common/assembly/impl/EthereumLikeWalletManagerAssembly.kt
@@ -1,22 +1,24 @@
-package com.tangem.blockchain.common.assembly
+package com.tangem.blockchain.common.assembly.impl
 
 import com.tangem.blockchain.blockchains.ethereum.EthereumTransactionBuilder
+import com.tangem.blockchain.blockchains.ethereum.EthereumWalletManager
 import com.tangem.blockchain.blockchains.ethereum.getEthereumJsonRpcProviders
 import com.tangem.blockchain.blockchains.ethereum.network.EthereumNetworkService
-import com.tangem.blockchain.blockchains.optimism.OptimismWalletManager
+import com.tangem.blockchain.common.assembly.WalletManagerAssembly
+import com.tangem.blockchain.common.assembly.WalletManagerAssemblyInput
 
-object OptimismWalletManagerAssembly : WalletManagerAssembly<OptimismWalletManager>() {
+object EthereumLikeWalletManagerAssembly : WalletManagerAssembly<EthereumWalletManager>() {
 
-    override fun make(input: WalletManagerAssemblyInput): OptimismWalletManager {
+    override fun make(input: WalletManagerAssemblyInput): EthereumWalletManager {
         with(input.wallet) {
-            return OptimismWalletManager(
+            return EthereumWalletManager(
                 wallet = this,
                 transactionBuilder = EthereumTransactionBuilder(
                     walletPublicKey = publicKey.blockchainKey,
                     blockchain = blockchain
                 ),
                 networkProvider = EthereumNetworkService(
-                    jsonRpcProviders = blockchain.getEthereumJsonRpcProviders(input.config),
+                    jsonRpcProviders = blockchain.getEthereumJsonRpcProviders(input.config)
                 ),
                 presetTokens = input.presetTokens
             )
