+++ blockchain/src/main/java/com/tangem/blockchain/common/assembly/impl/EthereumWalletManagerAssembly.kt
@@ -1,11 +1,14 @@
-package com.tangem.blockchain.common.assembly
+package com.tangem.blockchain.common.assembly.impl
 
 import com.tangem.blockchain.blockchains.ethereum.EthereumTransactionBuilder
 import com.tangem.blockchain.blockchains.ethereum.EthereumWalletManager
 import com.tangem.blockchain.blockchains.ethereum.getEthereumJsonRpcProviders
 import com.tangem.blockchain.blockchains.ethereum.network.EthereumNetworkService
+import com.tangem.blockchain.common.assembly.WalletManagerAssembly
+import com.tangem.blockchain.common.assembly.WalletManagerAssemblyInput
+import com.tangem.blockchain.network.blockcypher.BlockcypherNetworkProvider
 
-object EthereumLikeWalletManagerAssembly : WalletManagerAssembly<EthereumWalletManager>() {
+object EthereumWalletManagerAssembly : WalletManagerAssembly<EthereumWalletManager>() {
 
     override fun make(input: WalletManagerAssemblyInput): EthereumWalletManager {
         with(input.wallet) {
@@ -16,7 +19,11 @@ object EthereumLikeWalletManagerAssembly : WalletManagerAssembly<EthereumWalletM
                     blockchain = blockchain
                 ),
                 networkProvider = EthereumNetworkService(
-                    jsonRpcProviders = blockchain.getEthereumJsonRpcProviders(input.config)
+                    jsonRpcProviders = blockchain.getEthereumJsonRpcProviders(input.config),
+                    blockcypherNetworkProvider = BlockcypherNetworkProvider(
+                        blockchain = blockchain,
+                        tokens = input.config.blockcypherTokens
+                    )
                 ),
                 presetTokens = input.presetTokens
             )
