+++ blockchain/src/main/java/com/tangem/blockchain/common/assembly/impl/RavencoinWalletManagerAssembly.kt
@@ -1,9 +1,11 @@
-package com.tangem.blockchain.common.assembly
+package com.tangem.blockchain.common.assembly.impl
 
 import com.tangem.blockchain.blockchains.bitcoin.BitcoinTransactionBuilder
 import com.tangem.blockchain.blockchains.bitcoin.getBitcoinNetworkProviders
 import com.tangem.blockchain.blockchains.bitcoin.network.BitcoinNetworkService
 import com.tangem.blockchain.blockchains.ravencoin.RavencoinWalletManager
+import com.tangem.blockchain.common.assembly.WalletManagerAssembly
+import com.tangem.blockchain.common.assembly.WalletManagerAssemblyInput
 
 object RavencoinWalletManagerAssembly : WalletManagerAssembly<RavencoinWalletManager>() {
