+++ blockchain/src/main/java/com/tangem/blockchain/common/assembly/impl/SolanaWalletManagerAssembly.kt
@@ -1,7 +1,9 @@
-package com.tangem.blockchain.common.assembly
+package com.tangem.blockchain.common.assembly.impl
 
 import com.tangem.blockchain.blockchains.solana.SolanaRpcClientBuilder
 import com.tangem.blockchain.blockchains.solana.SolanaWalletManager
+import com.tangem.blockchain.common.assembly.WalletManagerAssembly
+import com.tangem.blockchain.common.assembly.WalletManagerAssemblyInput
 
 object SolanaWalletManagerAssembly : WalletManagerAssembly<SolanaWalletManager>() {
