+++ blockchain/src/main/java/com/tangem/blockchain/common/assembly/impl/TerraV1WalletManagerAssembly.kt
@@ -1,22 +1,24 @@
-package com.tangem.blockchain.common.assembly
+package com.tangem.blockchain.common.assembly.impl
 
 import com.tangem.blockchain.blockchains.cosmos.CosmosWalletManager
 import com.tangem.blockchain.blockchains.cosmos.network.CosmosChain
 import com.tangem.blockchain.blockchains.cosmos.network.CosmosRestProvider
+import com.tangem.blockchain.common.assembly.WalletManagerAssembly
+import com.tangem.blockchain.common.assembly.WalletManagerAssemblyInput
 import com.tangem.blockchain.extensions.letNotBlank
 
-object TerraV2WalletManagerAssembly : WalletManagerAssembly<CosmosWalletManager>() {
+object TerraV1WalletManagerAssembly : WalletManagerAssembly<CosmosWalletManager>() {
 
     override fun make(input: WalletManagerAssemblyInput): CosmosWalletManager {
         val providers = buildList {
-            input.config.getBlockCredentials?.apiKey.letNotBlank { add("https://luna.getblock.io/$it/mainnet/") }
-            add("https://phoenix-lcd.terra.dev/")
+            input.config.nowNodeCredentials?.apiKey.letNotBlank { add("https://terra.nownodes.io/$it/") }
+            add("https://terra-classic-lcd.publicnode.com/")
         }.map(::CosmosRestProvider)
 
         return CosmosWalletManager(
             wallet = input.wallet,
             networkProviders = providers,
-            cosmosChain = CosmosChain.TerraV2
+            cosmosChain = CosmosChain.TerraV1
         )
     }
