+++ blockchain/src/main/java/com/tangem/blockchain/common/assembly/impl/TonWalletManagerAssembly.kt
@@ -1,15 +1,9 @@
-package com.tangem.blockchain.common.assembly
+package com.tangem.blockchain.common.assembly.impl
 
 import com.tangem.blockchain.blockchains.ton.TonJsonRpcClientBuilder
 import com.tangem.blockchain.blockchains.ton.TonWalletManager
-import com.tangem.blockchain.blockchains.xrp.XrpTransactionBuilder
-import com.tangem.blockchain.blockchains.xrp.XrpWalletManager
-import com.tangem.blockchain.blockchains.xrp.network.XrpNetworkService
-import com.tangem.blockchain.blockchains.xrp.network.rippled.RippledNetworkProvider
-import com.tangem.blockchain.common.GetBlockCredentials
-import com.tangem.blockchain.common.NowNodeCredentials
-import com.tangem.blockchain.extensions.letNotBlank
-import com.tangem.blockchain.network.API_XRP_LEDGER_FOUNDATION
+import com.tangem.blockchain.common.assembly.WalletManagerAssembly
+import com.tangem.blockchain.common.assembly.WalletManagerAssemblyInput
 
 object TonWalletManagerAssembly : WalletManagerAssembly<TonWalletManager>() {
